test: move snapshot tests to dedicated file
